XRT-715 SonarQube fixes

+++ src/c/utests/json/json.c
@@ -28,12 +28,12 @@ static int dump_parse (const char *js, iot_json_tok_t *t, int count, int indent)
   }
   if (t->type == IOT_JSON_PRIMITIVE)
   {
-    printf ("%.*s", (int) (t->end - t->start), js + t->start);
+    printf ("%.*s", (t->end - t->start), js + t->start);
     return 1;
   }
   else if (t->type == IOT_JSON_STRING)
   {
-    printf ("'%.*s'", (int) (t->end - t->start), js + t->start);
+    printf ("'%.*s'", (t->end - t->start), js + t->start);
     return 1;
   }
   else if (t->type == IOT_JSON_OBJECT)
